+++ gms/api/src/main/snapshot/com.linkedin.identity.corpUsers.snapshot.json
@@ -338,7 +338,7 @@
     "namespace" : "com.linkedin.identity",
     "path" : "/corpUsers",
     "schema" : "com.linkedin.identity.CorpUser",
-    "doc" : "generated from: com.linkedin.identity.rest.resources.CorpUsers",
+    "doc" : "generated from: com.linkedin.metadata.resources.identity.CorpUsers",
     "collection" : {
       "identifier" : {
         "name" : "corpUser",
@@ -452,7 +452,7 @@
           "namespace" : "com.linkedin.identity",
           "path" : "/corpUsers/{corpUser}/editableInfo",
           "schema" : "com.linkedin.identity.CorpUserEditableInfo",
-          "doc" : "Rest.li entry point: /corpUsers/{corpUserKey}/editableInfo\n\ngenerated from: com.linkedin.identity.rest.resources.CorpUsersEditableInfoResource",
+          "doc" : "Rest.li entry point: /corpUsers/{corpUserKey}/editableInfo\n\ngenerated from: com.linkedin.metadata.resources.identity.CorpUsersEditableInfoResource",
           "collection" : {
             "identifier" : {
               "name" : "editableInfoId",
