+++ gms/impl/src/main/java/com/linkedin/metadata/resources/identity/CorpUsers.java
@@ -1,4 +1,4 @@
-package com.linkedin.identity.rest.resources;
+package com.linkedin.metadata.resources.identity;
 
 import com.linkedin.common.urn.CorpuserUrn;
 import com.linkedin.common.urn.Urn;
